Update and rename Standards to readme.md
Renaming the file like so will display it prominently on the repository front page, and give it some nice formatting with headings, paragraphs and highlighting the list items, etc.

I wasn't entirely sure what to put in the heading, but thought there should at least be some. If you have a better idea/want a different one, feel free to post that in the comment below or through code review, and I'll make the change before you pull it into the repo.